Updates and fixes
Deformer h was out of date. The subcharacters are not implemented and connected to deformer now. Bug fix where yellow was not yellow. Also blue is now 250 in b value. This is because i use blue as selected color for polygons. Shader recognizes this. Also when clicking on a material in Materials List, this color is applied on object level to selected objects.