Curve adding to objects improved
Still needs testing i think. If everything, including geo connections and memory management are fine, i can proceed to subdivision modifications in Subdivisions. There new logic need to be introduced that uses curves and segments to place point coordinates for subdivided geometry. Edges need segments in curves and vertexes need to scan for cps in curves. New center points are introduced for this in polys and quads also. Then we see if they are useful.