Curve deletion update
With cps verts are associated. When a curve is deleted, it may not delete cps, because other curves use them. So cps can use object vertexes to be associated with object. As object vertexes have edges surrounding them, the same way cps attached to geometry are surrounded with segments. If there is no segment in the edge, cp reflects it with NULL entry.