Since transition was jerky
Curve weight transition disabled, instead zero or one states inside transformed subdivision, at the end of it where vertices aquire curve coordinates for objects with curves. To get it working as intended too much needs to be changed and not certain, if it is possible then also. Since the shape of subdivided object uses its previous levels to determine transformed coordinates.